simulator/requirements.txt
web3==7.8.0
python-dotenv==1.0.1
rich==13.9.4
pyyaml==6.0.2
requests==2.32.3
